requirements.txt
Flask==2.3.3
twilio==8.11.0
python-dotenv==1.0.1
google-cloud-bigquery==3.34.0
